Added/improved some explicit casting to fix some warnings in the 32-bit MSVC build.
EXE_icons.cpp:291: warning C4244: 'argument': conversion from 'const off64_t' to 'const unsigned int', possible loss of data
TGA.cpp:315: warning C4244: 'initializing': conversion from 'off64_t' to 'size_t', possible loss of data
TGA.cpp:315: warning C4244: 'initializing': conversion from 'off64_t' to 'const size_t', possible loss of data

+++ src/libromdata/Other/EXE_icons.cpp
@@ -288,7 +288,7 @@ rp::uvector<uint8_t> EXEPrivate::loadIconResourceData(int iconindex, int width, 
 	}
 
 	rp::uvector<uint8_t> iconData;
-	iconData.resize(iconSize);
+	iconData.resize(static_cast<size_t>(iconSize));
 	size_t size = f_rtIcon->read(iconData.data(), iconData.size());
 	if (size != iconData.size()) {
 		// Read error.
